Refactor into Uniswap
